+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.1]
+
+### Added
+
+* Add command in group `Functional Tests` to run tests with option `--golden_run`.
+* Add setting parameter to use the host's `SCP_PRIVKEY` environment variable when loading/deleting app on device.
+* Add setting parameter to set the default device.
+
 ## [0.2.0]
 
 ### Added